Better touch event handling
Ignores all but the first pointer if multiple fingers are touching the
screen, and enables "touch-to-pan, pen-to-write" by default.

We are now also using scrollBy() to handle touch panning, which loses
the nice "momentum" of default panning, but saves us from a lot of the
jank of default panning (overshooting the scrolling boundaries,
cardinal axis snapping).