radio-panel.mjs: Added radio-panel component
A web component that treats its child button elements like a panel of
radio buttons: at most one may be pressed at a time, and the panel
element inherits the value of the pressed button.

The pressed button has the disabled attribute set. This is a bit of a
hack, but there are nebulous advantages to using standard attributes
of standard elements.